server/requirements.txt
fastapi
numpy
opencv-contrib-python
Pillow
uvicorn